+++ Frontend/src/components/interface.ts
@@ -1,0 +1,10 @@
+export interface User {
+  ID: number
+  Name: string
+  Gender: string
+  Address: string
+  Mobile: string
+  DOB: Date
+  Age: number
+  inEdit?: boolean | string
+}
