--- package.json
@@ -1,72 +1,0 @@
-{
-  "name": "ramp-up-phase-01",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@progress/kendo-data-query": "^1.6.0",
-    "@progress/kendo-drawing": "^1.17.2",
-    "@progress/kendo-licensing": "^1.3.0",
-    "@progress/kendo-react-animation": "^5.9.0",
-    "@progress/kendo-react-buttons": "^5.9.0",
-    "@progress/kendo-react-data-tools": "^5.9.0",
-    "@progress/kendo-react-dateinputs": "^5.9.0",
-    "@progress/kendo-react-dropdowns": "^5.9.0",
-    "@progress/kendo-react-grid": "^5.9.0",
-    "@progress/kendo-react-inputs": "^5.9.0",
-    "@progress/kendo-react-intl": "^5.9.0",
-    "@progress/kendo-react-treeview": "^5.9.0",
-    "@progress/kendo-theme-default": "^5.12.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.11",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.10",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "react-toastify": "^9.1.1",
-    "typescript": "^4.9.4",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint src/**/*.{js,jsx,ts,tsx,json}",
-    "lint:fix": "eslint --fix 'src/**/*.{js,jsx,ts,tsx,json}'",
-    "format": "prettier --write 'src/**/*.{js,jsx,ts,tsx,css,md,json}' --config ./.prettierrc"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.48.1",
-    "@typescript-eslint/parser": "^5.48.1",
-    "eslint": "^8.31.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.11",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.8.2"
-  }
-}
