--- src/components/interface.ts
@@ -1,10 +1,0 @@
-export interface User {
-  ID: number
-  Name: string
-  Gender: string
-  Address: string
-  Mobile: string
-  DOB: Date
-  Age: number
-  inEdit?: boolean | string
-}
