--- src/components/sample-data.tsx
@@ -1,29 +1,0 @@
-export const sampleData = [
-  {
-    ID: 1,
-    Name: 'Thilina Sandakelum',
-    Gender: 'Male',
-    Address: 'No. 1, 2nd Lane, Colombo 07',
-    Mobile: '0711234567',
-    DOB: new Date(1996, 8, 20),
-    Age: 27,
-  },
-  {
-    ID: 2,
-    Name: 'Kamal Nadeesha',
-    Gender: 'Male',
-    Address: 'No. 2, 2nd Lane, Colombo 07',
-    Mobile: '0711764567',
-    DOB: new Date(1996, 7, 20),
-    Age: 27,
-  },
-  {
-    ID: 3,
-    Name: 'Kushan Thathsara',
-    Gender: 'Male',
-    Address: 'No. 1, 4th Lane, Colombo 07',
-    Mobile: '0711290567',
-    DOB: new Date(1996, 2, 20),
-    Age: 27,
-  },
-]
